Rename schemas/announcements -> announcement
