+++ imports/lib/models/announcements.ts
@@ -1,9 +1,9 @@
 import { huntsMatchingCurrentUser } from '../../model-helpers';
-import AnnouncementsSchema, { AnnouncementType } from '../schemas/announcements';
+import AnnouncementSchema, { AnnouncementType } from '../schemas/announcement';
 import Base from './base';
 
 const Announcements = new Base<AnnouncementType>('announcements');
-Announcements.attachSchema(AnnouncementsSchema);
+Announcements.attachSchema(AnnouncementSchema);
 Announcements.publish(huntsMatchingCurrentUser);
 
 export default Announcements;
